p1-m10/requirements.txt
Flask==3.0.3
flask-swagger-ui==4.11.1
PyJWT==2.8.0
Werkzeug==3.0.2
psycopg2-binary==2.9.9
gunicorn==22.0.0